src/main/java/com/example/exam/service/ExaminerServiceImpl.java
package com.example.exam.service;

import com.example.exam.exception.BadRequestException;
import com.example.exam.model.Question;
import org.springframework.stereotype.Service;
import java.util.ArrayList;
import java.util.HashSet;
import java.util.List;
import java.util.Set;

@Service
public class ExaminerServiceImpl implements ExaminerService {
    private final QuestionService questionService;

    public ExaminerServiceImpl(QuestionService questionService) {
        this.questionService = questionService;
    }

    @Override
    public List<Question> getQuestions(int amount) {
        List<Question> allQuestions = questionService.getAllQuestions();
        if (amount > allQuestions.size()) {
            throw new BadRequestException("Запрошено больше вопросов, чем доступно.");
        }

        Set<Question> uniqueQuestions = new HashSet<>();
        while (uniqueQuestions.size() < amount) {
            uniqueQuestions.add(questionService.getRandomQuestion());
        }

        return new ArrayList<>(uniqueQuestions);
    }
}
